--- app/src/main/java/com/lateinit/rightweight/data/datasource/UserDataSource.kt
@@ -1,11 +1,0 @@
-package com.lateinit.rightweight.data.datasource
-
-import com.lateinit.rightweight.data.model.User
-import kotlinx.coroutines.flow.Flow
-
-interface UserDataSource {
-
-    suspend fun saveUser(user: User)
-
-    fun getUser(): Flow<User?>
-}
