+++ app/src/main/java/com/lateinit/rightweight/data/repository/UserRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.lateinit.rightweight.data.repository
 
-import com.lateinit.rightweight.data.datasource.UserDataSource
+import com.lateinit.rightweight.data.datasource.UserLocalDataSource
 import com.lateinit.rightweight.data.datasource.UserRemoteDataSource
 import com.lateinit.rightweight.data.model.User
 import com.lateinit.rightweight.data.model.WriteModelData
@@ -8,15 +8,15 @@ import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class UserRepositoryImpl @Inject constructor(
-    private val userDataSource: UserDataSource,
+    private val userLocalDataSource: UserLocalDataSource,
     private val userRemoteDataSource: UserRemoteDataSource
 ) : UserRepository {
     override suspend fun saveUser(user: User) {
-        userDataSource.saveUser(user)
+        userLocalDataSource.saveUser(user)
     }
 
     override fun getUser(): Flow<User?> {
-        return userDataSource.getUser()
+        return userLocalDataSource.getUser()
     }
 
     override suspend fun backupUserInfo(user: User) {
