+++ app/src/main/java/com/lateinit/rightweight/di/RepositoryModule.kt
@@ -32,7 +32,7 @@ class RepositoryModule {
     @Provides
     @Singleton
     fun getUserRepository(
-        userLocalDataSource: UserDataSource,
+        userLocalDataSource: UserLocalDataSource,
         userRemoteDataSource: UserRemoteDataSource
     ): UserRepository {
         return UserRepositoryImpl(userLocalDataSource, userRemoteDataSource)
